Finalizando

+++ atvi5/server/src/models/servicoTable.ts
@@ -15,11 +15,10 @@ const servicos = database.define('servicos', {
         allowNull: false
     },
     preco: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10,2),
         allowNull: false
     }
 })
-
 
 //servicos.sync({ alter: true });
